Move source to ./src Update test file path locations Update package.json
