+++ src/settings.js
